src/third_party/WebKit/LayoutTests/external/wpt/wasm/jsapi/module/customSections.any-expected.txt

This is a testharness.js-based test.
FAIL Missing arguments assert_throws: function "() => WebAssembly.Module.customSections(module)" did not throw
PASS Non-Module arguments
PASS Branding
PASS Empty module
PASS Empty module: array caching
FAIL Custom sections WebAssembly.Module(): Wasm decoding failed: unexpected section: name @+33
PASS Custom sections with surrogate pairs
FAIL Custom sections with U+FFFD assert_equals: expected 1 but got 0
Harness: the test ran to completion.
